# Add new `Performance/DeletePrefix` and `Performance/DeleteSuffix` cops