--- a/manual/cops_performance.md
+++ b/manual/cops_performance.md
@@ -249,6 +249,62 @@ Model.select('field AS field_one').count
 Model.select(:value).count
 ```
 
+## Performance/DeletePrefix
+
+Enabled by default | Safe | Supports autocorrection | VersionAdded | VersionChanged
+--- | --- | --- | --- | ---
+Enabled | Yes | Yes  | 1.6 | -
+
+In Ruby 2.5, `String#delete_prefix` has been added.
+
+This cop identifies places where `gsub(/\Aprefix/, '')`
+can be replaced by `delete_prefix('prefix')`.
+
+The `delete_prefix('prefix')` method is faster than
+`gsub(/\Aprefix/, '')`.
+
+### Examples
+
+```ruby
+# bad
+str.gsub(/\Aprefix/, '')
+str.gsub!(/\Aprefix/, '')
+str.gsub(/^prefix/, '')
+str.gsub!(/^prefix/, '')
+
+# good
+str.delete_prefix('prefix')
+str.delete_prefix!('prefix')
+```
+
+## Performance/DeleteSuffix
+
+Enabled by default | Safe | Supports autocorrection | VersionAdded | VersionChanged
+--- | --- | --- | --- | ---
+Enabled | Yes | Yes  | 1.6 | -
+
+In Ruby 2.5, `String#delete_suffix` has been added.
+
+This cop identifies places where `gsub(/suffix\z/, '')`
+can be replaced by `delete_suffix('suffix')`.
+
+The `delete_suffix('suffix')` method is faster than
+`gsub(/suffix\z/, '')`.
+
+### Examples
+
+```ruby
+# bad
+str.gsub(/suffix\z/, '')
+str.gsub!(/suffix\z/, '')
+str.gsub(/suffix$/, '')
+str.gsub!(/suffix$/, '')
+
+# good
+str.delete_suffix('suffix')
+str.delete_suffix!('suffix')
+```
+
 ## Performance/Detect
 
 Enabled by default | Safe | Supports autocorrection | VersionAdded | VersionChanged