@vishalzambre 2020-06-19
@koic  `delete_prefix` doesn't support to regex but rubocop gives error if gsub is used with regex