@koic 2020-06-19
@vishalzambre Can you open the issue with the following template?
https://github.com/rubocop-hq/rubocop-performance/issues/new?template=bug_report.md